+++ digid_eherkenning/saml2/eherkenning.py
@@ -423,7 +423,7 @@ class eHerkenningClient(BaseSaml2Client):
 
     @property
     def conf(self) -> dict:
-        if self._conf is None:
+        if not hasattr(self, "_conf"):
             db_config = EherkenningConfiguration.get_solo()
             self._conf = db_config.as_dict()
             self._conf.setdefault("acs_path", reverse("eherkenning:acs"))